@@ -7,8 +7,8 @@
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^0.27.2",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.4.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
@@ -18,12 +18,7 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "watch": "node-sass src/sass/main.scss src/style.css -w",
-    "compile:scss": "node-sass src/sass/main.scss src/style.css",
-    "add-prefix": "postcss --use autoprefixer -b 'last 10 versions' src/style.css -o src/styleWithPrifix.css",
-    "compress:css": "node-sass src/styleWithPrifix.css src/main.css --output-style compressed",
-    "build:css": "npm-run-all compile:scss add-prefix compress:css"
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
@@ -44,11 +39,8 @@
     ]
   },
   "devDependencies": {
-    "autoprefixer": "^10.4.7",
-    "node-sass": "^7.0.3",
-    "npm-run-all": "^4.1.5",
+    "autoprefixer": "^10.4.8",
     "postcss": "^8.4.14",
-    "postcss-cli": "^8.3.1",
-    "tailwindcss": "^3.1.2"
+    "tailwindcss": "^3.1.7"
   }
 }
