@@ -1,28 +1,34 @@
-import React, { Fragment } from "react";
-import { Route, Routes } from "react-router-dom";
-
-import Navbar from "./components/Navbar";
-import Home from "./components/Home";
-import SignIn from "./components/auth/SignIn";
-import SignUp from "./components/auth/SignUp";
+import React from "react";
+import { Routes, Route } from "react-router-dom";
+import Signin from "./components/auth/Signin";
+import Signup from "./components/auth/Signup";
 import EmailVerification from "./components/auth/EmailVerification";
-import ForgotPassword from "./components/auth/ForgotPassword";
+import Home from "./components/Home";
+import Navbar from "./components/user/Navbar";
+import ForgetPassword from "./components/auth/ForgetPassword";
 import ConfirmPassword from "./components/auth/ConfirmPassword";
 import NotFound from "./components/NotFound";
+import { useAuth } from "./hooks";
+import AdminNavigator from "./navigator/AdminNavigator";
 
 export default function App() {
+  const { authInfo } = useAuth();
+  const isAdmin = authInfo.profile?.role === "admin";
+
+  if (isAdmin) return <AdminNavigator />;
+
   return (
-    <Fragment>
+    <>
       <Navbar />
       <Routes>
         <Route path="/" element={<Home />} />
-        <Route path="/auth/sign-in" element={<SignIn />} />
-        <Route path="/auth/sign-up" element={<SignUp />} />
+        <Route path="/auth/signin" element={<Signin />} />
+        <Route path="/auth/signup" element={<Signup />} />
         <Route path="/auth/verification" element={<EmailVerification />} />
-        <Route path="/auth/forgot-password" element={<ForgotPassword />} />
+        <Route path="/auth/forget-password" element={<ForgetPassword />} />
         <Route path="/auth/reset-password" element={<ConfirmPassword />} />
         <Route path="*" element={<NotFound />} />
       </Routes>
-    </Fragment>
+    </>
   );
 }
