@@ -1,5 +1,7 @@
 import axios from "axios";
 
-const client = axios.create({ baseURL: "http://localhost:8000/api/v1" });
+const client = axios.create({
+  baseURL: "http://localhost:8000/api",
+});
 
 export default client;
