@@ -2,6 +2,6 @@ import React from "react";
 
 export default function Container({ children, className }) {
   return (
-    <div className={"max-w-screen-xl mx-auto" + className}>{children}</div>
+    <div className={"max-w-screen-xl mx-auto " + className}>{children}</div>
   );
 }
