@@ -1,7 +1,5 @@
-import React from 'react'
+import React from "react";
 
-export default function NotFound(){
-  return (
-    <div>NotFound </div>
-  )
+export default function NotFound() {
+  return <div>NotFound</div>;
 }
