@@ -1,118 +1,123 @@
 import React, { useEffect, useState } from "react";
-import { ImSpinner3 } from "react-icons/im";
 import { useNavigate, useSearchParams } from "react-router-dom";
-import { resetPassword, verifyPasswordResetToken } from "../../api/auth";
-import { useNotification } from "../../hooks/customHooks";
-import { commonModelClasses } from "../../utils/theme";
+import { ImSpinner3 } from "react-icons/im";
+import { commonModalClasses } from "../../utils/theme";
 import Container from "../Container";
-import { FormContainer } from "../form/FormContainer";
+import FormContainer from "../form/FormContainer";
 import FormInput from "../form/FormInput";
 import Submit from "../form/Submit";
-import Title from "./../form/Title";
+import Title from "../form/Title";
+import { resetPassword, verifyPasswordResetToken } from "../../api/auth";
+import { useNotification } from "../../hooks";
 
-export default function ForgotPassword() {
+export default function ConfirmPassword() {
   const [password, setPassword] = useState({
     one: "",
     two: "",
   });
-  const [isVerifying, setIsVerifying] = useState(true);
+  const [isVerifying, setIsVerifying] = useState(false);
   const [isValid, setIsValid] = useState(false);
   const [searchParams] = useSearchParams();
+  const token = searchParams.get("token");
+  const id = searchParams.get("id");
+
   const { updateNotification } = useNotification();
   const navigate = useNavigate();
-  const token = searchParams.get("token");
-  const userId = searchParams.get("id");
+
+  // isValid, !isValid
+  useEffect(() => {
+    isValidToken();
+  }, []);
 
   const isValidToken = async () => {
-    const { error, valid } = await verifyPasswordResetToken(token, userId);
+    const { error, valid } = await verifyPasswordResetToken(token, id);
     setIsVerifying(false);
     if (error) {
-      navigate("/auth/reset-password", {
-        replace: true,
-      });
+      navigate("/auth/reset-password", { replace: true });
       return updateNotification("error", error);
     }
+
     if (!valid) {
       setIsValid(false);
-      return navigate("/auth/reset-password", {
-        replace: true,
-      });
+      return navigate("/auth/reset-password", { replace: true });
     }
+
     setIsValid(true);
   };
 
-  useEffect(() => {
-    isValidToken();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-  if (isVerifying) {
-    return (
-      <FormContainer>
-        <Container>
-          <div className=" flex space-x-2 items-center">
-            <h1 className=" text-4xl font-semibold dark:text-white text-primary">
-              Please wait while we are verifying Your token
-            </h1>
-            <ImSpinner3 className=" animate-spin  text-4xl dark:text-white text-primary" />
-          </div>
-        </Container>
-      </FormContainer>
-    );
-  }
-  if (!isValid) {
-    return (
-      <FormContainer>
-        <Container>
-          <div className=" flex space-x-2 items-center">
-            <h1 className=" text-4xl font-semibold dark:text-white text-primary">
-              Sorry, Your token is not valid
-            </h1>
-          </div>
-        </Container>
-      </FormContainer>
-    );
-  }
-  const handleChange = ({target}) => {
-    const {name, value} = target;
-     setPassword({...password, [name]:value});
+  const handleChange = ({ target }) => {
+    const { name, value } = target;
+    setPassword({ ...password, [name]: value });
   };
-  const handleSubmit = async(e) => {
+
+  const handleSubmit = async (e) => {
     e.preventDefault();
-    if(!password.one.trim()){
-      return updateNotification('error', 'Password is missing!')
-    }
-    if(password.one.trim().length < 8){
-      return updateNotification('error', 'Password must be 8 characters long! ')
-    }
-    if(password.one !== password.two){
-      return updateNotification('error', 'Password not matched!')
-    }
-    const {error, message} = await resetPassword({newPassword: password.one, userId, token})
-    if(error){
-      return updateNotification('error', error);
-    } 
-    updateNotification('success', message);
-    navigate('/auth/sign-in', {replace: true}); 
-  }
+
+    if (!password.one.trim())
+      return updateNotification("error", "Password is missing!");
+
+    if (password.one.trim().length < 8)
+      return updateNotification("error", "Password must be 8 characters long!");
+
+    if (password.one !== password.two)
+      return updateNotification("error", "Password do not match!");
+
+    const { error, message } = await resetPassword({
+      newPassword: password.one,
+      userId: id,
+      token,
+    });
+
+    if (error) return updateNotification("error", error);
+
+    updateNotification("success", message);
+    navigate("/auth/signin", { replace: true });
+  };
+
+  if (isVerifying)
+    return (
+      <FormContainer>
+        <Container>
+          <div className="flex space-x-2 items-center">
+            <h1 className="text-4xl font-semibold dark:text-white text-primary">
+              Please wait we are verifying your token!
+            </h1>
+            <ImSpinner3 className="animate-spin text-4xl dark:text-white text-primary" />
+          </div>
+        </Container>
+      </FormContainer>
+    );
+
+  if (!isValid)
+    return (
+      <FormContainer>
+        <Container>
+          <h1 className="text-4xl font-semibold dark:text-white text-primary">
+            Sorry the token is invalid!
+          </h1>
+        </Container>
+      </FormContainer>
+    );
+
   return (
     <FormContainer>
       <Container>
-        <form onSubmit={handleSubmit} className={commonModelClasses + " w-96"}>
+        <form onSubmit={handleSubmit} className={commonModalClasses + " w-96"}>
           <Title>Enter New Password</Title>
           <FormInput
-            label="New Password"
-            placeholder="***************"
-            name="one"
             value={password.one}
             onChange={handleChange}
+            label="New Password"
+            placeholder="********"
+            name="one"
             type="password"
           />
           <FormInput
-            label="Confirm Password"
-            placeholder="***************"
-            name="two"
-            onChange={handleChange}
             value={password.two}
+            onChange={handleChange}
+            label="Confirm Password"
+            placeholder="********"
+            name="two"
             type="password"
           />
           <Submit value="Confirm Password" />
