@@ -1,75 +1,87 @@
-import React, { useState, useEffect, useRef } from "react";
+import React, { useEffect, useRef, useState } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 import { resendEmailVerificationToken, verifyUserEmail } from "../../api/auth";
-import { useAuth, useNotification } from "../../hooks/customHooks";
-import { commonModelClasses } from "../../utils/theme";
+import { useAuth, useNotification } from "../../hooks";
+import { commonModalClasses } from "../../utils/theme";
 import Container from "../Container";
-import { FormContainer } from "../form/FormContainer";
+import FormContainer from "../form/FormContainer";
 import Submit from "../form/Submit";
-import Title from "./../form/Title";
+import Title from "../form/Title";
 
-const OPT_LENGTH = 6;
+const OTP_LENGTH = 6;
 let currentOTPIndex;
+
 const isValidOTP = (otp) => {
+  let valid = false;
+
   for (let val of otp) {
-    if (isNaN(parseInt(val))) return false;
+    valid = !isNaN(parseInt(val));
+    if (!valid) break;
   }
-  return true;
+
+  return valid;
 };
+
 export default function EmailVerification() {
-  const [otp, setOpt] = useState(new Array(OPT_LENGTH).fill(""));
+  const [otp, setOtp] = useState(new Array(OTP_LENGTH).fill(""));
   const [activeOtpIndex, setActiveOtpIndex] = useState(0);
 
-  const inputRef = useRef();
-  const navigate = useNavigate();
-  const { updateNotification } = useNotification();
-  const { state } = useLocation();
   const { isAuth, authInfo } = useAuth();
   const { isLoggedIn, profile } = authInfo;
   const isVerified = profile?.isVerified;
+
+  const inputRef = useRef();
+  const { updateNotification } = useNotification();
+
+  const { state } = useLocation();
   const user = state?.user;
-  useEffect(() => {
-    inputRef.current?.focus();
-  }, [activeOtpIndex]);
-  useEffect(() => {
-    if (!user) navigate("/not-found");
-    if (isLoggedIn && isVerified) navigate("/");
-  }, [user, isLoggedIn, isVerified ,navigate]);
+
+  const navigate = useNavigate();
 
   const focusNextInputField = (index) => {
     setActiveOtpIndex(index + 1);
   };
+
   const focusPrevInputField = (index) => {
-    index = index === 0 ? 0 : index - 1;
-    setActiveOtpIndex(index);
+    let nextIndex;
+    const diff = index - 1;
+    nextIndex = diff !== 0 ? diff : 0;
+    setActiveOtpIndex(nextIndex);
   };
-  const handleOtpChange = (e) => {
-    const value = e.target.value;
-    let newOtp = [...otp];
+
+  const handleOtpChange = ({ target }) => {
+    const { value } = target;
+    const newOtp = [...otp];
     newOtp[currentOTPIndex] = value.substring(value.length - 1, value.length);
-    if (!value) {
-      focusPrevInputField(currentOTPIndex);
-    } else {
-      focusNextInputField(currentOTPIndex);
-    }
-    setOpt([...newOtp]);
+
+    if (!value) focusPrevInputField(currentOTPIndex);
+    else focusNextInputField(currentOTPIndex);
+    setOtp([...newOtp]);
   };
+
+  const handleOTPResend = async () => {
+    const { error, message } = await resendEmailVerificationToken(user.id);
+
+    if (error) return updateNotification("error", error);
+
+    updateNotification("success", message);
+  };
+
   const handleKeyDown = ({ key }, index) => {
     currentOTPIndex = index;
     if (key === "Backspace") {
       focusPrevInputField(currentOTPIndex);
     }
   };
-  const handleOtpResend = async() => {
-    const {error, message} = await resendEmailVerificationToken(user.id)
-    if(error) return updateNotification('error', error);
-    updateNotification('success', message);
-  }
+
   const handleSubmit = async (e) => {
     e.preventDefault();
+
     if (!isValidOTP(otp)) {
-      return updateNotification("error", "Invalid OTP!!");
+      return updateNotification("error", "invalid OTP");
     }
+
+    // submit otp
     const {
       error,
       message,
@@ -79,41 +91,58 @@ export default function EmailVerification() {
       userId: user.id,
     });
     if (error) return updateNotification("error", error);
+
     updateNotification("success", message);
-    console.log(userResponse.token);
     localStorage.setItem("auth-token", userResponse.token);
     isAuth();
-    navigate("/");
   };
+
+  useEffect(() => {
+    inputRef.current?.focus();
+  }, [activeOtpIndex]);
+
+  useEffect(() => {
+    if (!user) navigate("/not-found");
+    if (isLoggedIn && isVerified) navigate("/");
+  }, [user, isLoggedIn, isVerified]);
+
+  // if(!user) return null
 
   return (
     <FormContainer>
       <Container>
-        <form onSubmit={handleSubmit} className={commonModelClasses}>
+        <form onSubmit={handleSubmit} className={commonModalClasses}>
           <div>
-            <Title>Please Enter OPT to verify your account</Title>
+            <Title>Please enter the OTP to verify your account</Title>
             <p className="text-center dark:text-dark-subtle text-light-subtle">
-              OTP has been sent to your Email
+              OTP has been sent to your email
             </p>
           </div>
-          <div className=" flex justify-center items-center space-x-4">
+
+          <div className="flex justify-center items-center space-x-4">
             {otp.map((_, index) => {
               return (
                 <input
                   ref={activeOtpIndex === index ? inputRef : null}
                   key={index}
-                  className=" w-12 h-12 border-2 rounded dark:border-dark-subtle dark:focus:border-white focus:border-primary bg-transparent outline-none text-center dark:text-white text-primary font-semibold text-xl spin-button-none"
                   type="number"
                   value={otp[index] || ""}
-                  onChange={(e) => handleOtpChange(e, index)}
+                  onChange={handleOtpChange}
                   onKeyDown={(e) => handleKeyDown(e, index)}
+                  className="w-12 h-12 border-2 dark:border-dark-subtle  border-light-subtle darK:focus:border-white focus:border-primary rounded bg-transparent outline-none text-center dark:text-white text-primary font-semibold text-xl spin-button-none"
                 />
               );
             })}
           </div>
           <div>
-          <Submit value="Verify Account" />
-          <button onClick={handleOtpResend} type="button" className="dark:text-white text-blue-500 font-semibold hover:underline mt-2">I don't have OTP</button>
+            <Submit value="Verify Account" />
+            <button
+              onClick={handleOTPResend}
+              type="button"
+              className="dark:text-white text-blue-500 font-semibold hover:underline mt-2"
+            >
+              I don't have OTP
+            </button>
           </div>
         </form>
       </Container>
