@@ -1,50 +1,52 @@
 import React, { useState } from "react";
-import { commonModelClasses } from "../../utils/theme";
+import { forgetPassword } from "../../api/auth";
+import { useNotification } from "../../hooks";
+import { isValidEmail } from "../../utils/helper";
+import { commonModalClasses } from "../../utils/theme";
 import Container from "../Container";
 import CustomLink from "../CustomLink";
-import { FormContainer } from "../form/FormContainer";
+import FormContainer from "../form/FormContainer";
 import FormInput from "../form/FormInput";
 import Submit from "../form/Submit";
-import Title from "./../form/Title";
-import { isValidEmail } from "../../utils/helper";
-import { useNotification } from "../../hooks/customHooks";
-import { forgetPassword } from "../../api/auth";
+import Title from "../form/Title";
 
-export default function ForgotPassword() {
+export default function ForgetPassword() {
   const [email, setEmail] = useState("");
   const { updateNotification } = useNotification();
+
   const handleChange = ({ target }) => {
     const { value } = target;
     setEmail(value);
   };
+
   const handleSubmit = async (e) => {
     e.preventDefault();
-    if (!isValidEmail(email)) {
-      return updateNotification("error", "Invalid Email! ");
-    }
+    if (!isValidEmail(email))
+      return updateNotification("error", "Invalid email!");
+
     const { error, message } = await forgetPassword(email);
     if (error) return updateNotification("error", error);
-    updateNotification("success", message);
+
+    updateNotification("error", message);
   };
+
   return (
     <FormContainer>
       <Container>
-        <form
-          onSubmit={handleSubmit}
-          className={commonModelClasses + "w-96 space-y-6"}
-        >
+        <form onSubmit={handleSubmit} className={commonModalClasses + " w-96"}>
           <Title>Please Enter Your Email</Title>
           <FormInput
             onChange={handleChange}
             value={email}
             label="Email"
-            placeholder="abcd@gmail.com"
+            placeholder="john@email.com"
             name="email"
           />
           <Submit value="Send Link" />
+
           <div className="flex justify-between">
-            <CustomLink to="/auth/sign-in">Sign in </CustomLink>
-            <CustomLink to="/auth/sign-up">Sign up</CustomLink>
+            <CustomLink to="/auth/signin">Sign in</CustomLink>
+            <CustomLink to="/auth/signup">Sign up</CustomLink>
           </div>
         </form>
       </Container>
