@@ -1,73 +1,80 @@
-import React, { useState, useEffect } from "react";
+import React, { useEffect, useState } from "react";
+import { useNavigate } from "react-router-dom";
+import { useAuth, useNotification } from "../../hooks";
+import { isValidEmail } from "../../utils/helper";
+import { commonModalClasses } from "../../utils/theme";
 import Container from "../Container";
 import CustomLink from "../CustomLink";
+import FormContainer from "../form/FormContainer";
 import FormInput from "../form/FormInput";
 import Submit from "../form/Submit";
-import Title from "./../form/Title";
-import { commonModelClasses } from "../../utils/theme";
-import { FormContainer } from "../form/FormContainer";
-import { useAuth, useNotification } from "../../hooks/customHooks";
-import { useNavigate } from "react-router-dom";
-import { isValidEmail } from "../../utils/helper";
+import Title from "../form/Title";
 
 const validateUserInfo = ({ email, password }) => {
-  if (!isValidEmail(email)) return { ok: false, error: "Invalid Email!!" };
-  if (!password.trim()) return { ok: false, error: "Password is missing!!" };
+  if (!email.trim()) return { ok: false, error: "Email is missing!" };
+  if (!isValidEmail(email)) return { ok: false, error: "Invalid email!" };
+
+  if (!password.trim()) return { ok: false, error: "Password is missing!" };
   if (password.length < 8)
-    return { ok: false, error: "Password must be 8 characters long!! " };
-  return { ok: true, error: null };
+    return { ok: false, error: "Password must be 8 characters long!" };
+
+  return { ok: true };
 };
 
-export default function SignIn() {
-  const [userInfo, setUserinfo] = useState({
-    name: "",
+export default function Signin() {
+  const [userInfo, setUserInfo] = useState({
     email: "",
     password: "",
   });
+
   const navigate = useNavigate();
   const { updateNotification } = useNotification();
   const { handleLogin, authInfo } = useAuth();
   const { isPending, isLoggedIn } = authInfo;
 
-  useEffect(() => {
-    if (isLoggedIn) navigate("/");
-  }, [isLoggedIn, navigate]);
-
   const handleChange = ({ target }) => {
     const { value, name } = target;
-    setUserinfo({ ...userInfo, [name]: value });
+    setUserInfo({ ...userInfo, [name]: value });
   };
+
   const handleSubmit = async (e) => {
     e.preventDefault();
     const { ok, error } = validateUserInfo(userInfo);
+
     if (!ok) return updateNotification("error", error);
     handleLogin(userInfo.email, userInfo.password);
   };
 
+  // useEffect(() => {
+  //   // we want to move our user to somewhere else
+  //   if (isLoggedIn) navigate("/");
+  // }, [isLoggedIn]);
+
   return (
     <FormContainer>
       <Container>
-        <form onSubmit={handleSubmit} className={commonModelClasses + " w-72"}>
+        <form onSubmit={handleSubmit} className={commonModalClasses + " w-72"}>
           <Title>Sign in</Title>
           <FormInput
             value={userInfo.email}
             onChange={handleChange}
             label="Email"
-            placeholder="abcd@gmail.com"
+            placeholder="john@email.com"
             name="email"
           />
           <FormInput
             value={userInfo.password}
             onChange={handleChange}
             label="Password"
-            placeholder="**************"
+            placeholder="********"
             name="password"
             type="password"
           />
           <Submit value="Sign in" busy={isPending} />
+
           <div className="flex justify-between">
-            <CustomLink to="/auth/forgot-password">Forgot Password</CustomLink>
-            <CustomLink to="/auth/sign-up">Sign up</CustomLink>
+            <CustomLink to="/auth/forget-password">Forget password</CustomLink>
+            <CustomLink to="/auth/signup">Sign up</CustomLink>
           </div>
         </form>
       </Container>
