@@ -1,6 +1,7 @@
 import React, { createContext, useEffect, useState } from "react";
+import { useNavigate } from "react-router-dom";
 import { getIsAuth, signInUser } from "../api/auth";
-import { useNotification } from "../hooks/customHooks";
+import { useNotification } from "../hooks";
 
 export const AuthContext = createContext();
 
@@ -11,25 +12,31 @@ const defaultAuthInfo = {
   error: "",
 };
 
-export const AuthProvider = ({ children }) => {
+export default function AuthProvider({ children }) {
   const [authInfo, setAuthInfo] = useState({ ...defaultAuthInfo });
-
   const { updateNotification } = useNotification();
+
+  const navigate = useNavigate();
+
   const handleLogin = async (email, password) => {
     setAuthInfo({ ...authInfo, isPending: true });
     const { error, user } = await signInUser({ email, password });
     if (error) {
-      updateNotification('error', error);
+      updateNotification("error", error);
       return setAuthInfo({ ...authInfo, isPending: false, error });
     }
+
+    navigate("/", { replace: true });
     setAuthInfo({
       profile: { ...user },
-      isLoggedIn: true,
       isPending: false,
+      isLoggedIn: true,
       error: "",
     });
+
     localStorage.setItem("auth-token", user.token);
   };
+
   const isAuth = async () => {
     const token = localStorage.getItem("auth-token");
     if (!token) return;
@@ -37,8 +44,10 @@ export const AuthProvider = ({ children }) => {
     setAuthInfo({ ...authInfo, isPending: true });
     const { error, user } = await getIsAuth(token);
     if (error) {
+      updateNotification("error", error);
       return setAuthInfo({ ...authInfo, isPending: false, error });
     }
+
     setAuthInfo({
       profile: { ...user },
       isLoggedIn: true,
@@ -46,23 +55,22 @@ export const AuthProvider = ({ children }) => {
       error: "",
     });
   };
+
   const handleLogout = () => {
     localStorage.removeItem("auth-token");
-    console.log("hi");
     setAuthInfo({ ...defaultAuthInfo });
   };
 
   useEffect(() => {
     isAuth();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
+  //  handleLogout
   return (
     <AuthContext.Provider
-      value={{ authInfo, handleLogin, isAuth, handleLogout }}
+      value={{ authInfo, handleLogin, handleLogout, isAuth }}
     >
-      {" "}
-      {children}{" "}
+      {children}
     </AuthContext.Provider>
   );
-};
+}
