@@ -5,33 +5,32 @@ export const ThemeContext = createContext();
 const defaultTheme = "light";
 const darkTheme = "dark";
 
-export const ThemeProvider = ({ children }) => {
+export default function ThemeProvider({ children }) {
   const toggleTheme = () => {
     const oldTheme = getTheme();
     const newTheme = oldTheme === defaultTheme ? darkTheme : defaultTheme;
+
     updateTheme(newTheme, oldTheme);
   };
+
   useEffect(() => {
     const theme = getTheme();
-    if (!theme) {
-      updateTheme(defaultTheme)
-    } else {
-      updateTheme(theme);
-    }
+    if (!theme) updateTheme(defaultTheme);
+    else updateTheme(theme);
   }, []);
+
   return (
     <ThemeContext.Provider value={{ toggleTheme }}>
       {children}
     </ThemeContext.Provider>
   );
-};
+}
 
-// Themes
-const getTheme = () => localStorage.getItem('theme');
+const getTheme = () => localStorage.getItem("theme");
+
 const updateTheme = (theme, themeToRemove) => {
-  if(themeToRemove){  
-    document.documentElement.classList.remove(themeToRemove);
-  }
+  if (themeToRemove) document.documentElement.classList.remove(themeToRemove);
+
   document.documentElement.classList.add(theme);
   localStorage.setItem("theme", theme);
-}
+};
