@@ -1,7 +1,7 @@
-import { useContext } from "react"
+import { useContext } from "react";
 import { AuthContext } from "../context/AuthProvider";
 import { NotificationContext } from "../context/NotificationProvider";
-import { ThemeContext } from "../context/ThemeProvider"
+import { ThemeContext } from "../context/ThemeProvider";
 
 export const useTheme = () => useContext(ThemeContext);
 export const useNotification = () => useContext(NotificationContext);
