@@ -1,4 +1,5 @@
 export const isValidEmail = (email) => {
-  const isValid = /^\w+([.-]?\w+)*@\w+([.-]?\w+)*(\.\w{2,3})+$/;
+  const isValid = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
+
   return isValid.test(email);
 };
