@@ -1,1 +1,2 @@
-export const commonModelClasses = "dark:bg-secondary bg-white drop-shadow-lg rounded p-6 space-y-6"
+export const commonModalClasses =
+  "dark:bg-secondary bg-white drop-shadow-lg rounded p-6 space-y-6";
