@@ -1,3 +1,4 @@
+/** @type {import('tailwindcss').Config} */
 module.exports = {
   content: ["./src/**/*.{js,jsx}"],
   darkMode: "class",
@@ -8,8 +9,6 @@ module.exports = {
         secondary: "#272727",
         "dark-subtle": "rgba(255, 255, 255, 0.5)",
         "light-subtle": "rgba(39, 39, 39, 0.5)",
-        "highlight-dark": "#ffc200",
-        highlight: "#058bfb",
       },
     },
   },
